--- chat.json
@@ -1,40 +1,0 @@
-{
-  "name": "My Conversation",
-  "messages": [
-    {
-      "content": "Hello there!",
-      "timestamp": 1448470901,
-      "senderId": "bob"
-    },
-    {
-      "content": "how are you?",
-      "timestamp": 1448470905,
-      "senderId": "mike"
-    },
-    {
-      "content": "I\u0027m good thanks, do you like pie?",
-      "timestamp": 1448470906,
-      "senderId": "bob"
-    },
-    {
-      "content": "no, let me ask Angus...",
-      "timestamp": 1448470910,
-      "senderId": "mike"
-    },
-    {
-      "content": "Hell yes! Are we buying some pie?",
-      "timestamp": 1448470912,
-      "senderId": "angus"
-    },
-    {
-      "content": "No, just want to know if there\u0027s anybody else in the pie society...",
-      "timestamp": 1448470914,
-      "senderId": "bob"
-    },
-    {
-      "content": "YES! I\u0027m the head pie eater there...",
-      "timestamp": 1448470915,
-      "senderId": "angus"
-    }
-  ]
-}
