+++ logs/app-info-2021-02-16.log
@@ -1,0 +1,7 @@
+[INFO ] 2021-02-16 21:52:04.509 [main] ConversationExporter - Conversation exported from 'chat.txt' to 'chat.json
+[INFO ] 2021-02-16 21:52:07.445 [main] ConversationExporter - Conversation exported from 'chat.txt' to 'chat.json
+[INFO ] 2021-02-16 21:52:17.694 [main] ConversationExporter - Conversation exported from 'chat.txt' to 'chat.json
+[INFO ] 2021-02-16 21:52:39.713 [main] ConversationExporter - Conversation exported from 'chat.txt' to 'chat.json
+[INFO ] 2021-02-16 21:52:43.774 [main] ConversationExporter - Conversation exported from 'chat.txt' to 'chat.json
+[INFO ] 2021-02-16 21:52:54.863 [main] ConversationExporter - Conversation exported from 'chat.txt' to 'chat.json
+[INFO ] 2021-02-16 21:53:03.350 [main] ConversationExporter - Conversation exported from 'chat.txt' to 'chat.json
