+++ src/main/java/com/mindlinksoft/recruitment/mychat/ConversationExporterConfiguration.java
@@ -1,13 +1,16 @@
 package com.mindlinksoft.recruitment.mychat;
 
+// import com.mindlinksoft.recruitment.mychat.models.Conversation;
+// import com.mindlinksoft.recruitment.mychat.models.Message;
+// import com.mindlinksoft.recruitment.mychat.options.*;
+
 import picocli.CommandLine.Option;
 import picocli.CommandLine.Command;
 
 /**
  * Represents the configuration for the exporter.
  */
-@Command(name = "export", mixinStandardHelpOptions = true, version = "exporter 1.0",
-         description = "Exports a plain text chat log into a JSON file.")
+@Command(name = "export", mixinStandardHelpOptions = true, version = "exporter 1.0", description = "Exports a plain text chat log into a JSON file.")
 public final class ConversationExporterConfiguration {
     /**
      * Gets the input file path.
