+++ src/main/java/com/mindlinksoft/recruitment/mychat/models/Conversation.java
@@ -1,4 +1,4 @@
-package com.mindlinksoft.recruitment.mychat;
+package com.mindlinksoft.recruitment.mychat.models;
 
 import java.util.Collection;
 
@@ -18,7 +18,8 @@ public final class Conversation {
 
     /**
      * Initializes a new instance of the {@link Conversation} class.
-     * @param name The name of the conversation.
+     * 
+     * @param name     The name of the conversation.
      * @param messages The messages in the conversation.
      */
     public Conversation(String name, Collection<Message> messages) {
