+++ src/main/java/com/mindlinksoft/recruitment/mychat/models/Message.java
@@ -1,4 +1,4 @@
-package com.mindlinksoft.recruitment.mychat;
+package com.mindlinksoft.recruitment.mychat.models;
 
 import java.time.Instant;
 
@@ -23,9 +23,10 @@ public final class Message {
 
     /**
      * Initializes a new instance of the {@link Message} class.
+     * 
      * @param timestamp The timestamp at which the message was sent.
-     * @param senderId The ID of the sender.
-     * @param content The message content.
+     * @param senderId  The ID of the sender.
+     * @param content   The message content.
      */
     public Message(Instant timestamp, String senderId, String content) {
         this.content = content;
